+++ src/ElectromagneticFields.jl
@@ -44,4 +44,8 @@ include("analytic/singular.jl")
 include("analytic/symmetric_quadratic.jl")
 include("analytic/theta_pinch.jl")
 
+export plot_equilibrium, plot_equilibrium!
+
+include("plots.jl")
+
 end