+++ src/analytic/abc.jl
@@ -10,7 +10,6 @@ Parameters: `a`, `b`, `c`
 """
 module ABC
 
-    using RecipesBase
 
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
@@ -54,49 +53,9 @@ module ABC
     ElectromagneticFields.A₁(x::AbstractVector, equ::ABCEquilibrium) = equ.a₀ * sin(x[3]) + equ.c₀ * cos(x[2])
     ElectromagneticFields.A₂(x::AbstractVector, equ::ABCEquilibrium) = equ.b₀ * sin(x[1]) + equ.a₀ * cos(x[3])
     ElectromagneticFields.A₃(x::AbstractVector, equ::ABCEquilibrium) = equ.c₀ * sin(x[2]) + equ.b₀ * cos(x[1])
-    B(x::AbstractVector, equ::ABCEquilibrium) = A₁(x,equ)^2 + A₂(x,equ)^2 + A₃(x,equ)^2
+    B(x::AbstractVector, equ::ABCEquilibrium) = sqrt(A₁(x,equ)^2 + A₂(x,equ)^2 + A₃(x,equ)^2)
 
     ElectromagneticFields.get_functions(::ABCEquilibrium) = (X=X, Y=Y, Z=Z)
 
-
-    @recipe function f(equ::ABCEquilibrium; nx=99, ni=div(nx,2)+1, nl=12, size=(400,1200))
-        xmin = 0
-        xmax = 2π
-        grid = LinRange(xmin, xmax, nx)
-        Bfield = [B([grid[i], grid[j], grid[k]], equ) for i in eachindex(grid), j in eachindex(grid), k in eachindex(grid)]
-
-        seriestype   := :contour
-        aspect_ratio := :equal
-        layout := (3, 1)
-        size   := size
-        xlims  := (xmin,xmax)
-        ylims  := (xmin,xmax)
-        levels := nl
-        legend := :none
-
-        @series begin
-            subplot := 1
-            xguide := "x"
-            yguide := "y"
-            title  := "|B(x,y,π)|"
-            (grid, grid, Bfield[:,:,ni])
-        end
-
-        @series begin
-            subplot := 2
-            xguide := "x"
-            yguide := "z"
-            title  := "|B(x,π,z)|"
-            (grid, grid, Bfield[:,ni,:])
-        end
-
-        @series begin
-            subplot := 3
-            xguide := "y"
-            yguide := "z"
-            title  := "|B(π,y,z)|"
-            (grid, grid, Bfield[ni,:,:])
-        end
-    end
 
 end