+++ src/analytic/penning_trap_uniform.jl
@@ -26,7 +26,6 @@ Parameters:
 """
 module PenningTrapUniform
 
-    using RecipesBase
 
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments