+++ src/analytic/solovev_symmetric.jl
@@ -14,7 +14,6 @@ Parameters:
 """
 module SolovevSymmetric
 
-    using RecipesBase
 
     import ..ElectromagneticFields
     import ..ElectromagneticFields: CartesianEquilibrium, code, code_arguments
@@ -67,26 +66,4 @@ module SolovevSymmetric
     ElectromagneticFields.get_functions(::SolovevSymmetricEquilibrium) = (X=X, Y=Y, Z=Z)
 
 
-    @recipe function f(equ::SolovevSymmetricEquilibrium;
-                       nx = 100, ny = 120, levels = 25, size = (600,400),
-                       xlims = (equ.R₀-0.75, equ.R₀+0.75),
-                       ylims = (-0.50, +0.50))
-
-        xgrid = LinRange(xlims[1], xlims[2], nx)
-        zgrid = LinRange(ylims[1], ylims[2], ny)
-        pot   = [ElectromagneticFields.A₃([xgrid[i], zgrid[j], 0.0], equ) for i in eachindex(xgrid), j in eachindex(zgrid)]
-
-        seriestype := :contour
-        aspect_ratio := :equal
-        size   := size
-        xlims  := xlims
-        ylims  := ylims
-        levels := levels
-        legend := :none
-        xguide := "x"
-        yguide := "y"
-
-        (xgrid, zgrid, pot')
-    end
-    
 end